Restructured source code to maven standard directory layout and improved maven script to build a standalone jar file without dependencies.
