+++ src/main/java/com/jhlabs/image/AbstractBufferedImageOp.java
