+++ src/main/java/com/jhlabs/image/ConvolveFilter.java
