+++ src/main/java/com/jhlabs/image/GaussianFilter.java
