+++ src/main/java/com/jhlabs/image/ImageMath.java
