+++ src/main/java/wagner/stephanie/lizzie/rules/BoardHistoryNode.java
