+++ src/main/java/wagner/stephanie/lizzie/rules/Stone.java
